+++ genmonmaint.sh
@@ -96,6 +96,7 @@ function updatelibraries() {
   else
     sudo $pipcommand install pyotp==2.3.0 -U
   fi
+  sudo $pipcommand install mopeka_pro_check -U
   echo "Done."
 }
 
@@ -164,6 +165,7 @@ function installgenmon() {
     fi
     # correct problem with LDAP3 module install
     sudo $pipcommand install pyasn1 -U
+    sudo $pipcommand install mopeka_pro_check 
 
     sudo chmod 775 "$genmondir/startgenmon.sh"
     sudo chmod 775 "$genmondir/genmonmaint.sh"